plug-ins: add extra layer groups when loading PSD images with clipping layers
Together with the previous commit this mostly fixes #5438.

PhotoShop handles clipping layers in a different way than GIMP. The only
way to have it look the same is by adding extra layer groups.
PSD layers that have clipping set, combined with the first non clipping
layer below it are grouped together in a new layer group.
Doing this results in the same image as the PSD merged image unless there
are other PSD elements in play that we don't handle yet.

See e.g. the NSFW image in the mentioned issue where the purple hair color
is a little darker than the merged image.