plug-ins: fix #5305 Lighting Effects crashes when I use a channel as Bumpmap
When a channel is selected as bumpmap, the bump_format can be different
than it was before. However, we always kept the first bump_format, which
could be RGB, in which case trying to use it with a channel caused a
crash in GIMP.

To fix this we always update the bump_format if bumpmap is enabled.

(cherry picked from commit 7b7d61635891ef7f7d6aee6f8eec3e372c93b8c5)

# Conflicts:
#	plug-ins/lighting/lighting-image.c