.gitlab: search common git ancestor with `mr-origin` remote.
It looks like origin is the same as mr-origin when the contributor
pushes to one's branch. But when a reviewer rebases through the Gitlab
"Rebase" button on web GUI, I got a fatal error:

> fatal: ambiguous argument 'origin/asalle/use-dynamics-flag': unknown
  revision or path not in the working tree.

Possibly `origin` is then the remote of the person who rebased (it would
be weird, yet it's a fact the branch is not found). Let's go with this
assumption.