INSTALL: update the install instructions regarding librsvg.
Currently it's a mandatory option (and it has been the case for years,
ever since commit 43e218859b09) so let's update the info.

Note that there are still discussions going on about this dependency and
it being hard or impossible to build on many platforms (which are stuck
on old C version, before the move to Rust). See #6821.
We'll see how it goes.