Issue #7720: Layer Sets UI Suggestion.
Using the "list-add" icon instead of "document-save" indeed makes more
sense for the button to store a layer set.

Also renaming "Named Selection" to "New layer set's name", which should
hopefully be less confusing. In particular, we keep consistency with the
"No layer set stored" label (naming the concept "layer set" both times).
And it should be clearer that it is a field to write down a name for
creating a new layer set.